# openapi.yaml
...
openapi: 3.1.0
info:
  title: Code Titans — PS8 + PS2 Question Trust Pipeline
  description: "\n**Code Titans — HackWithAMYPO 2026 Stage 2 MVP**\n\nTwo problem\
    \ statements wired into one pipeline:\n\n* **PS8** — Assignment Question Iteration\
    \ & Variation Generation\n* **PS2** — AI Hallucination Detection & Reliability\
    \ Scoring\n\n```\nseed -> parse -> plan -> generate (Qwen2.5-Coder 7B via Ollama)\n\
    \     -> PS8 structural validation -> PS2 reliability verification\n     -> decision\
    \ -> PASS | REVIEW | REJECT -> (regenerate on REJECT)\n```\n\nEverything runs\
    \ locally. No paid third-party APIs are used anywhere.\n"
  version: 1.0.0
paths:
...
  /api/v1/progress/{job_id}:
    get:
      tags:
      - Integrated
      summary: Live progress of a generate-and-verify run
      operationId: get_progress_api_v1_progress__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Progress Api V1 Progress  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/generate-and-verify:
    post:
      tags:
      - Integrated
      summary: 'Run the full pipeline: generate -> validate -> verify -> decide ->
        regenerate'
      description: Execute the complete Code Titans trust pipeline for one seed question.
      operationId: generate_and_verify_api_v1_generate_and_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAndVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAndVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/demo/run:
    post:
      tags:
      - Integrated
      summary: 'DEMO: deterministic reject -> regenerate -> pass run of the real pipeline'
      description: 'Replay the scripted demo candidates through the real validation
        pipeline.


        Only the model call is scripted; every score, verdict and decision is computed

        live. Results are labelled DEMO and never written to any store.'
      operationId: run_demo_api_v1_demo_run_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DemoRunRequest'
              - type: 'null'
              title: Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAndVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/questions:
    get:
      tags:
      - Storage
      summary: Accepted question bank
      operationId: get_questions_api_v1_questions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Questions Api V1 Questions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/review-queue:
    get:
      tags:
      - Storage
      summary: Questions awaiting human review
      operationId: get_review_queue_api_v1_review_queue_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Review Queue Api V1 Review Queue Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/review-queue/{item_id}/approve:
    post:
      tags:
      - Storage
      summary: Approve a review item into the bank
      operationId: approve_review_item_api_v1_review_queue__item_id__approve_post
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ReviewAction'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Approve Review Item Api V1 Review Queue  Item Id  Approve
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/review-queue/{item_id}/reject:
    post:
      tags:
      - Storage
      summary: Reject a review item
      operationId: reject_review_item_api_v1_review_queue__item_id__reject_post
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ReviewAction'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reject Review Item Api V1 Review Queue  Item Id  Reject
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/validation-reports:
    get:
      tags:
      - Storage
      summary: Combined PS8 + PS2 validation reports
      operationId: get_validation_reports_api_v1_validation_reports_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Validation Reports Api V1 Validation Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/audit-logs:
    get:
      tags:
      - Storage
      summary: Audit trail
      operationId: get_audit_logs_api_v1_audit_logs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Audit Logs Api V1 Audit Logs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/export:
    get:
      tags:
      - Storage
      summary: Export the question bank as JSON or CSV
      description: Bulk export module (PS8 mandatory component).
      operationId: export_bank_api_v1_export_get
      parameters:
      - name: fmt
        in: query
        required: false
        schema:
          type: string
          pattern: ^(json|csv)$
          default: json
          title: Fmt
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/jobs:
    post:
      tags:
      - Jobs
      summary: Start a background generate-and-verify job
      operationId: start_job_api_v1_jobs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Start Job Api V1 Jobs Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Jobs
      summary: Recent jobs (without candidate results)
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Jobs Api V1 Jobs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - Jobs
      summary: Live state of one job, including candidate results
      operationId: get_job_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Job Api V1 Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Candidate:
      properties:
        id:
          type: string
          title: Id
        question:
          type: string
          title: Question
        answer_key:
          type: string
          title: Answer Key
        domain:
          type: string
          title: Domain
        topic:
          type: string
          title: Topic
        subtopic:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtopic
        difficulty:
          type: string
          enum:
          - easy
          - medium
          - hard
          - expert
          title: Difficulty
        difficulty_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Difficulty Score
          default: 0.5
        question_type:
          type: string
          title: Question Type
        learning_objective:
          type: string
          title: Learning Objective
        test_cases:
          items:
            $ref: '#/components/schemas/TestCase'
          type: array
          title: Test Cases
        variation_strategy:
          type: string
          title: Variation Strategy
        strategy_label:
          type: string
          title: Strategy Label
          default: ''
        solution_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Solution Method
        attempt:
          type: integer
          title: Attempt
          default: 0
        regenerated:
          type: boolean
          title: Regenerated
          default: false
        raw_model_output:
          anyOf:
          - type: string
          - type: 'null'
          title: Raw Model Output
        parse_warnings:
          items:
            type: string
          type: array
          title: Parse Warnings
      type: object
      required:
      - id
      - question
      - answer_key
      - domain
      - topic
      - difficulty
      - question_type
      - learning_objective
      - variation_strategy
      title: Candidate
      description: A generated question candidate before (or after) validation.
...
    DemoRunRequest:
      properties:
        job_id:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9_-]{1,64}$
          - type: 'null'
          title: Job Id
      type: object
      title: DemoRunRequest
...
    GenerateAndVerifyRequest:
      properties:
        seed_question:
          type: string
          minLength: 5
          title: Seed Question
        domain:
          type: string
          title: Domain
          default: programming
        count:
          type: integer
          maximum: 60.0
          minimum: 1.0
          title: Count
          default: 5
        difficulty_shift:
          anyOf:
          - type: string
            enum:
            - easy
            - medium
            - hard
            - expert
          - type: 'null'
          title: Difficulty Shift
        enable_regeneration:
          type: boolean
          title: Enable Regeneration
          default: true
        persist:
          type: boolean
          title: Persist
          description: Write accepted/review results to local JSON stores.
          default: true
        job_id:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9_-]{1,64}$
          - type: 'null'
          title: Job Id
          description: Optional client-chosen id; poll GET /progress/{job_id} while
            it runs.
      type: object
      required:
      - seed_question
      title: GenerateAndVerifyRequest
    GenerateAndVerifyResponse:
      properties:
        seed_metadata:
          anyOf:
          - $ref: '#/components/schemas/SeedMetadata'
          - type: 'null'
        results:
          items:
            $ref: '#/components/schemas/PipelineCandidate'
          type: array
          title: Results
        summary:
          $ref: '#/components/schemas/PipelineSummary'
        timings_ms:
          additionalProperties:
            type: number
          type: object
          title: Timings Ms
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        demo:
          type: boolean
          title: Demo
          default: false
        demo_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Demo Label
        demo_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Demo Description
      type: object
      required:
      - results
      - summary
      title: GenerateAndVerifyResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobRequest:
      properties:
        seed_question:
          type: string
          minLength: 5
          title: Seed Question
        domain:
          type: string
          title: Domain
          default: programming
        subject_area:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject Area
          description: Taxonomy subject area; overrides `domain` with its pipeline
            domain.
        count:
          type: integer
          maximum: 60.0
          minimum: 1.0
          title: Count
          default: 10
        difficulty_shift:
          anyOf:
          - type: string
            enum:
            - easy
            - medium
            - hard
            - expert
          - type: 'null'
          title: Difficulty Shift
        enable_regeneration:
          type: boolean
          title: Enable Regeneration
          default: true
        persist:
          type: boolean
          title: Persist
          default: true
      type: object
      required:
      - seed_question
      title: JobRequest
      description: Start a background generate-and-verify job (POST /jobs).
    PipelineCandidate:
      properties:
        candidate:
          $ref: '#/components/schemas/Candidate'
        decision:
          type: string
          enum:
          - PASS
          - REVIEW
          - REJECT
          title: Decision
        decision_reasons:
          items:
            type: string
          type: array
          title: Decision Reasons
        structural_validation:
          anyOf:
          - $ref: '#/components/schemas/StructuralValidation'
          - type: 'null'
        reliability_verification:
          anyOf:
          - $ref: '#/components/schemas/ReliabilityVerification'
          - type: 'null'
        status:
          type: string
          enum:
          - generated
          - validated
          - verified
          - accepted
          - rejected
          - review
          title: Status
          default: generated
        attempts:
          type: integer
          title: Attempts
          default: 1
        regeneration_history:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Regeneration History
        subject_area:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject Area
      type: object
      required:
      - candidate
      - decision
      title: PipelineCandidate
      description: A candidate plus everything both gates concluded about it.
    PipelineSummary:
      properties:
        requested:
          type: integer
          title: Requested
        generated:
          type: integer
          title: Generated
        passed:
          type: integer
          title: Passed
        review:
          type: integer
          title: Review
        rejected:
          type: integer
          title: Rejected
        duplicate_rate:
          type: number
          title: Duplicate Rate
        regeneration_attempts:
          type: integer
          title: Regeneration Attempts
        mean_reliability_score:
          type: number
          title: Mean Reliability Score
        flagged_span_count:
          type: integer
          title: Flagged Span Count
      type: object
      required:
      - requested
      - generated
      - passed
      - review
      - rejected
      - duplicate_rate
      - regeneration_attempts
      - mean_reliability_score
      - flagged_span_count
      title: PipelineSummary
    ReliabilityVerification:
      properties:
        reliability_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Reliability Score
        confidence_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence Score
        hallucination_probability:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Hallucination Probability
        verdict:
          type: string
          enum:
          - trustworthy
          - partially_reliable
          - misleading
          - fabricated
          - unverifiable
          title: Verdict
        flagged_spans:
          items:
            $ref: '#/components/schemas/FlaggedSpan'
          type: array
          title: Flagged Spans
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceItem'
          type: array
          title: Evidence
        claims:
          items:
            $ref: '#/components/schemas/Claim'
          type: array
          title: Claims
        signals:
          additionalProperties:
            type: number
          type: object
          title: Signals
        contradictions:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Contradictions
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        recommendations:
          items:
            type: string
          type: array
          title: Recommendations
        embedding_backend:
          type: string
          title: Embedding Backend
          default: ''
        self_consistency_used:
          type: boolean
          title: Self Consistency Used
          default: false
        timings_ms:
          additionalProperties:
            type: number
          type: object
          title: Timings Ms
      type: object
      required:
      - reliability_score
      - confidence_score
      - hallucination_probability
      - verdict
      title: ReliabilityVerification
    ReviewAction:
      properties:
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
      type: object
      title: ReviewAction
    SeedMetadata:
      properties:
        domain:
          type: string
          title: Domain
        topic:
          type: string
          title: Topic
        subtopic:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtopic
        core_concept:
          type: string
          title: Core Concept
        question_type:
          type: string
          title: Question Type
        difficulty:
          type: string
          enum:
          - easy
          - medium
          - hard
          - expert
          title: Difficulty
        difficulty_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Difficulty Score
        learning_objective:
          type: string
          title: Learning Objective
        keywords:
          items:
            type: string
          type: array
          title: Keywords
        raw_seed:
          type: string
          title: Raw Seed
      type: object
      required:
      - domain
      - topic
      - core_concept
      - question_type
      - difficulty
      - difficulty_score
      - learning_objective
      - raw_seed
      title: SeedMetadata
      description: What the Seed Parser extracts from a raw seed question.
    StructuralValidation:
      properties:
        concept_preserved:
          type: boolean
          title: Concept Preserved
        difficulty_match:
          type: boolean
          title: Difficulty Match
        is_duplicate:
          type: boolean
          title: Is Duplicate
        meaningful_variation:
          type: boolean
          title: Meaningful Variation
        passed:
          type: boolean
          title: Passed
        semantic_similarity:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Semantic Similarity
        lexical_similarity:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Lexical Similarity
        concept_overlap:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Concept Overlap
        difficulty_delta:
          type: number
          title: Difficulty Delta
        nearest_match:
          anyOf:
          - type: string
          - type: 'null'
          title: Nearest Match
        nearest_match_source:
          anyOf:
          - type: string
          - type: 'null'
          title: Nearest Match Source
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        checks:
          additionalProperties: true
          type: object
          title: Checks
      type: object
      required:
      - concept_preserved
      - difficulty_match
      - is_duplicate
      - meaningful_variation
      - passed
      - semantic_similarity
      - lexical_similarity
      - concept_overlap
      - difficulty_delta
      title: StructuralValidation
    TestCase:
      properties:
        input:
          type: string
          title: Input
          default: ''
        expected_output:
          type: string
          title: Expected Output
          default: ''
      type: object
      title: TestCase
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...